+++ all_receitas/lib/menu_principal.dart
@@ -1,6 +1,15 @@
 import 'package:flutter/material.dart';
 import 'var_json.dart';
 
+import 'receita1/parte2/receita1_2.dart' as rec1_2;
+
+
+
+void _receita2page(BuildContext context) {
+  Navigator.  of(context).push(
+    MaterialPageRoute(builder: (context) => rec1_2.Receita1Parte2()),
+  );
+}
 
 class MenuPrincipal extends StatelessWidget {
   const MenuPrincipal({super.key});
@@ -23,12 +32,18 @@ class MenuPrincipal extends StatelessWidget {
                     child: Wrap(
                       spacing: 20,
                       runSpacing: 20,
-                      children: menuItens.map((item) {
-                        return CustomCard(
-                          title: item['title'],
-                          rota: item['rota'],
-                        );
-                      }).toList(),
+                      children: [
+                        ElevatedButton(
+                          onPressed: () => _receita2page(context),
+                          child: Text("Receita 1_2")
+                        ),
+                      ],
+                      // children: menuItens.map((item) {
+                      //   return CustomCard(
+                      //     title: item['title'],
+                      //     rota: item['rota'],
+                      //   );
+                      // }).toList(),
                     ),
                   ),
                 )
